Rename dummy_close_loop to example_newsletter_signup everywere
