--- lib/talkable/generators/templates/app/views/dummy_close_loop/register.html.erb
@@ -1,4 +1,0 @@
-<%= form_tag :dummy_close_loop do %>
-  <%= email_field_tag :email %>
-  <%= submit_tag %>
-<% end %>
