+++ lib/talkable/generators/templates/app/views/example_newsletter_signup/register.html.erb
@@ -1,0 +1,4 @@
+<%= form_tag :example_newsletter_signup do %>
+  <%= email_field_tag :email %>
+  <%= submit_tag %>
+<% end %>
