+++ lib/talkable/generators/templates/app/views/example_newsletter_signup/register.html.haml
@@ -1,0 +1,3 @@
+= form_tag :example_newsletter_signup do
+  = email_field_tag :email
+  = submit_tag
