+++ lib/talkable/generators/templates/app/views/example_newsletter_signup/thank_you.html.erb
